fix: Arreglando nombre de paquetes y composables
- Los nombres de los composables se definen comenzando con mayuscula, el propio Android Studio te lo dice
- Los nombres de los paquetes van el minuscula, el propio Android Studio te lo dice
