+++ app/src/main/java/com/daviddev/passwordmanager/MainActivity.kt
@@ -7,22 +7,14 @@ import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
 import androidx.activity.viewModels
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.lifecycle.ViewModel
-import com.daviddev.passwordmanager.DeepSeekCode.AccountScreen
-import com.daviddev.passwordmanager.ViewModels.AddAccountViewModel
-import com.daviddev.passwordmanager.Views.AccountManagementScreen
-import com.daviddev.passwordmanager.Views.AddDataView
-import com.daviddev.passwordmanager.Views.PasswordManagerView
-import com.daviddev.passwordmanager.Views.dropdowntest
+import com.daviddev.passwordmanager.viewmodels.AddAccountViewModel
 import com.daviddev.passwordmanager.ui.theme.PasswordManagerTheme
 import dagger.hilt.android.AndroidEntryPoint
-import com.daviddev.passwordmanager.Views.test2
 import java.time.LocalDateTime
 
 @AndroidEntryPoint
@@ -37,7 +29,7 @@ class MainActivity : ComponentActivity() {
             PasswordManagerTheme{
                 val pairList = listOf(Pair("lolo", "123"))
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    com.daviddev.passwordmanager.Views.AddDataView();
+                    com.daviddev.passwordmanager.views.AddDataView();
                     //com.daviddev.passwordmanager.Views.test2()
                     Log.i("Seguimiento","Main activity dentro del scaffold" )
                 }
