--- app/src/main/java/com/daviddev/passwordmanager/ViewModels/PasswordManagerViewModel.kt
@@ -1,5 +1,0 @@
-package com.daviddev.passwordmanager.ViewModels
-
-class PasswordManagerViewModel {
-
-}
