--- app/src/main/java/com/daviddev/passwordmanager/Views/enternamedialog.kt
@@ -1,4 +1,0 @@
-package com.daviddev.passwordmanager.Views
-
-class EnterNameDialog {
-}
