+++ app/src/main/java/com/daviddev/passwordmanager/room/AccountDataDataBase.kt
@@ -1,14 +1,11 @@
-package com.daviddev.passwordmanager.Room
+package com.daviddev.passwordmanager.room
 //com.example.myapplication.Room.AccountDataBase.kt
 
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
-import com.daviddev.passwordmanager.Room.AccountDataDatabaseDao
-import com.daviddev.passwordmanager.Room.AccountDataDataBase
 import com.daviddev.passwordmanager.Model.KeyValueAccountData
-import com.daviddev.passwordmanager.Room.Converter
 
 @Database(entities = [KeyValueAccountData::class], version = 1, exportSchema = false)
 @TypeConverters(Converter::class)
