+++ app/src/main/java/com/daviddev/passwordmanager/room/DatabaseDAO.kt
@@ -1,4 +1,4 @@
-package com.daviddev.passwordmanager.Room
+package com.daviddev.passwordmanager.room
 //
 //com.example.myapplication.Room.AccountDataRepository.kt
 //com.example.myapplication.Room.AccountDataBase.kt
