+++ app/src/main/java/com/daviddev/passwordmanager/viewmodels/PasswordManagerViewModel.kt
@@ -1,0 +1,5 @@
+package com.daviddev.passwordmanager.viewmodels
+
+class PasswordManagerViewModel {
+
+}
