+++ app/src/main/java/com/daviddev/passwordmanager/views/AccountManagementScreen.kt
@@ -1,4 +1,4 @@
-package com.daviddev.passwordmanager.Views
+package com.daviddev.passwordmanager.views
 
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
@@ -9,7 +9,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.graphics.Color
 //import androidx.compose.material.icons.Icons
 //import androidx.compose.material.icons.filled.DateRange
 //import androidx.compose.material.icons.filled.Add
@@ -17,7 +16,6 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.ui.text.font.FontWeight
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
-import java.time.temporal.ChronoUnit
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
