+++ app/src/main/java/com/daviddev/passwordmanager/views/DatePicker.kt
@@ -1,4 +1,4 @@
-package com.daviddev.passwordmanager.Views
+package com.daviddev.passwordmanager.views
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
