+++ app/src/main/java/com/daviddev/passwordmanager/views/Enternamedialog.kt
@@ -1,0 +1,4 @@
+package com.daviddev.passwordmanager.views
+
+class EnterNameDialog {
+}
